+++ src/front/js/component/navbar.js
@@ -10,9 +10,6 @@ export const Navbar = () => {
 			<Link to="/signup">
 				<span className="navbar-brand mb-0 h1">Sign Up</span>
 			</Link>
-			<Link to="/signUpPage">
-				<span className="navbar-brand mb-0 h1">Sign Up Page</span>
-			</Link>
 			<Link to="/login">
 				<span className="navbar-brand mb-0 h1">Log In</span>
 			</Link>
