+++ src/front/js/layout.js
@@ -6,7 +6,6 @@ import { Home } from "./pages/home";
 import { Demo } from "./pages/demo";
 import { Single } from "./pages/single";
 import { SignUp } from "./pages/signup";
-import { SignUpPage } from "./pages/signuppage";
 import { LogIn } from "./pages/login";
 import injectContext from "./store/appContext";
 
@@ -14,6 +13,8 @@ import { Navbar } from "./component/navbar";
 import { Footer } from "./component/footer";
 import { Profile } from "./pages/profile";
 import { RegisterOrganization } from "./pages/registerOrganization";
+import { ToastContainer } from "react-toastify";
+import "react-toastify/dist/ReactToastify.css";
 
 //create your first component
 const Layout = () => {
@@ -26,6 +27,7 @@ const Layout = () => {
 			<BrowserRouter basename={basename}>
 				<ScrollToTop>
 					<Navbar />
+					<ToastContainer />
 					<Switch>
 						<Route exact path="/">
 							<Home />
@@ -35,9 +37,6 @@ const Layout = () => {
 						</Route>
 						<Route exact path="/signup">
 							<SignUp />
-						</Route>
-						<Route exact path="/signUpPage">
-							<SignUpPage />
 						</Route>
 						<Route exact path="/login">
 							<LogIn />
