+++ src/front/js/pages/signup.js
@@ -1,79 +1,156 @@
 import React, { useState } from "react";
+import { Link, useHistory } from "react-router-dom";
+import PropTypes from "prop-types";
 
-export const SignUp = () => {
-	const [email, setEmail] = useState("");
-	const [password, setPassword] = useState("");
-	const [confirmPassword, setConfirmPassword] = useState("");
-	const [error, setError] = useState("");
-	const [message, setMessage] = useState("");
+//validacines y notificaciones
+import SimpleReactValidator from "simple-react-validator";
+import { toast } from "react-toastify";
 
-	function singup() {
-		setError("");
-		if (password != confirmPassword) {
-			setError("Las contraseñas no coinciden");
-			return;
-		}
+//vistas
+import Grid from "@material-ui/core/Grid";
+import TextField from "@material-ui/core/TextField";
+import Button from "@material-ui/core/Button";
 
-		let responseOk = false;
-		fetch(process.env.BACKEND_URL + "/api/v1/signup", {
-			method: "POST",
-			headers: {
-				"Content-Type": "application/json"
-			},
-			body: JSON.stringify({
-				email: email,
-				password: password
-			})
+import "../../styles/signup.scss";
+
+export const SignUp = props => {
+	const history = useHistory();
+	const [value, setValue] = useState({
+		email: "",
+		password: "",
+		confirm_password: ""
+	});
+	const changeHandler = e => {
+		setValue({ ...value, [e.target.name]: e.target.value });
+		validator.showMessages();
+	};
+
+	const [validator] = React.useState(
+		new SimpleReactValidator({
+			className: "errorMessage"
 		})
-			.then(response => {
-				responseOk = response.ok;
-				if (response.ok) {
-					if (response.status === 201) {
-						setMessage("Usuario registrado correctamente");
-					} else if (response.status === 202) {
-						setMessage("Este usuario ya se encuentra registrado");
+	);
+
+	const submitForm = e => {
+		e.preventDefault();
+		if (validator.allValid()) {
+			let responseOk = false;
+			fetch(process.env.BACKEND_URL + "/api/v1/signup", {
+				method: "POST",
+				headers: {
+					"Content-Type": "application/json"
+				},
+				body: JSON.stringify({
+					email: value.email,
+					password: value.password
+				})
+			})
+				.then(response => {
+					responseOk = response.ok;
+					if (response.ok) {
+						if (response.status === 201) {
+							setValue({
+								email: "",
+								password: "",
+								confirm_password: ""
+							});
+							validator.hideMessages();
+							toast.success("Usuario registrado correctamente");
+							history.push("/login");
+						} else if (response.status === 202) {
+							toast.warn("Este usuario ya se encuentra registrado");
+						}
 					}
-				}
-				return response.json();
-			})
-			.then(responseJson => {
-				console.log(responseJson);
-				if (!responseOk) {
-					setError(responseJson.message);
-				}
-			})
-			.catch(error => {
-				console.log("error", error);
-				setError(error.message);
-			});
-	}
+					return response.json();
+				})
+				.then(responseJson => {
+					console.log(responseJson);
+					if (!responseOk) {
+						toast.error(responseJson.message);
+					}
+				})
+				.catch(error => {
+					console.log("error", error);
+					toast.error(error.message);
+				});
+		} else {
+			validator.showMessages();
+			toast.error("Empty field is not allowed!");
+		}
+	};
 
 	return (
-		<div className="jumbotron">
-			{error ? <h3>{error}</h3> : ""}
-			{message ? <h3>{message}</h3> : ""}
-			<input
-				type="email"
-				placeholder="Enter email"
-				onChange={event => {
-					setEmail(event.target.value);
-				}}
-			/>
-			<input
-				type="password"
-				placeholder="Enter password"
-				onChange={event => {
-					setPassword(event.target.value);
-				}}
-			/>
-			<input
-				type="password"
-				placeholder="Confirm password"
-				onChange={event => {
-					setConfirmPassword(event.target.value);
-				}}
-			/>
-			<input type="button" value="Crear" onClick={singup} />
-		</div>
+		<Grid className="loginWrapper">
+			<Grid className="loginForm">
+				<h2>Signup</h2>
+				<p>Signup your account</p>
+				<form onSubmit={submitForm}>
+					<Grid container spacing={3}>
+						<Grid item xs={12}>
+							<TextField
+								className="inputOutline"
+								fullWidth
+								placeholder="E-mail"
+								value={value.email}
+								variant="outlined"
+								name="email"
+								label="E-mail"
+								InputLabelProps={{
+									shrink: true
+								}}
+								onBlur={e => changeHandler(e)}
+								onChange={e => changeHandler(e)}
+							/>
+							{validator.message("email", value.email, "required|email")}
+						</Grid>
+						<Grid item xs={12}>
+							<TextField
+								className="inputOutline"
+								fullWidth
+								placeholder="Password"
+								value={value.password}
+								variant="outlined"
+								name="password"
+								label="Password"
+								InputLabelProps={{
+									shrink: true
+								}}
+								onBlur={e => changeHandler(e)}
+								onChange={e => changeHandler(e)}
+							/>
+							{validator.message("password", value.password, "required")}
+						</Grid>
+						<Grid item xs={12}>
+							<TextField
+								className="inputOutline"
+								fullWidth
+								placeholder="Confirm Password"
+								value={value.password}
+								variant="outlined"
+								name="confirm_password"
+								label="Confirm Password"
+								InputLabelProps={{
+									shrink: true
+								}}
+								onBlur={e => changeHandler(e)}
+								onChange={e => changeHandler(e)}
+							/>
+							{validator.message("confirm password", value.confirm_password, `in:${value.password}`)}
+						</Grid>
+						<Grid item xs={12}>
+							<Grid className="formFooter">
+								<Button fullWidth className="cBtn cBtnLarge cBtnTheme" type="submit">
+									Sign Up
+								</Button>
+							</Grid>
+
+							<p className="noteHelp">
+								Already have an account? <Link to="/login">Return to Login In</Link>
+							</p>
+						</Grid>
+					</Grid>
+				</form>
+			</Grid>
+		</Grid>
 	);
 };
